flytt funksjonalitet til egen service
